Adding source code
I've got some of the base design implemented, I hope to have the shaders
load from an assets directory. I'd like to implement OpenGL 2.0 and
Vulkan support. Maybe an osmesa backend could be implemented which
renders to a temporary file. Soon, I'll add objects and textures but I'd
like to do more with shaders first. I might add support for disabling
the tests in the CMakeLists and add in automatic compiling of GTest if
its not found.